pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "hockey-server"
version = "0.1.0"
description = "Hockey scoreboard FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "sqlalchemy[asyncio]>=2.0",
  "aiosqlite>=0.20",
  "pydantic-settings>=2.6",
  "bcrypt>=4.2",
  "pyjwt>=2.10",
  "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "httpx>=0.27"]

[tool.setuptools.packages.find]
where = ["."]
include = ["hockey_server*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
